[IMP] impl function delete 30%

+++ back/main.py
@@ -407,7 +407,7 @@ def list_fiche_metier():
 
 
 @verif_db
-@app.route("/api/v1/delete_content/", methods=['DELETE'])
+@app.route("/api/v1/delete_content", methods=['DELETE'])
 @jwt_required()
 def delete_contents():
     """
@@ -451,7 +451,7 @@ def delete_contents():
 
 
 @verif_db
-@app.route("/api/v1/delete_account/", methods=['DELETE'])
+@app.route("/api/v1/delete_account", methods=['DELETE'])
 @jwt_required()
 def delete_account():
     """
@@ -460,16 +460,24 @@ def delete_account():
     """
     try:
         access = get_jwt_identity().split("+")[1]
-        account_id = request.get_json().get("account_id")
+        data = request.get_json()
+        print(data)
+        if data:
+            account_id = data.get("compte_id")
+        else:
+            return {
+                "error": True,
+                "message": "Pas de compte à suprimmé"
+            }, 400
 
-        if access == "ADMIN":
+        if access == "ADMIN" and account_id:
             try:
                 CURSOR.execute("""
                     DELETE FROM
                         Compte
                     WHERE
                         id=%s;
-                """, (account_id))
+                """, (account_id,))
 
                 DB.commit()
 
@@ -479,7 +487,6 @@ def delete_account():
                 }, 200
             except Exception as err:
                 print(err)
-                DB.close()
                 return {
                     "error": True,
                     "message": "le compte est encore utilisé !"
@@ -496,7 +503,7 @@ def delete_account():
 
 
 @verif_db
-@app.route("/api/v1/delete_fiche_metier/", methods=['DELETE'])
+@app.route("/api/v1/delete_fiche_metier", methods=['DELETE'])
 @jwt_required()
 def delete_fiche_metier():
     """
